Store Student ID in long long instead of int
This fixes an overflow issue when having a 10 digit ID. The reason this
is long long instead of just long is for cross-compatibility. If I
recall correctly, long and int are the same size on Windows. And, since
I haven't switched to using things like int64_t, I'll just set it to
long long.

+++ data.h
@@ -50,12 +50,12 @@ enum class Answer
 struct Info
 {
     int thread_id = 0;
-    int id = 0;
+    long long id = 0;
     std::vector<Answer> answers;
 
     Info() { }
     Info(int t) :thread_id(t) { }
-    Info(int t, int i, const std::vector<Answer>& answers)
+    Info(int t, long long i, const std::vector<Answer>& answers)
         :thread_id(t), id(i), answers(answers) { }
 };
